Add files via upload
These have been my first two small projects that I have created with Python in a long time. I have created a random password generator and a safe access system. These two can go hand in hand together to make a more customised experience for users when attempting to make a password, The random password works well as in todays society, password cracking is something that is large and Cybersecurity is an important asset of life today.