Collecting data for a monitoring Dash
In this code, I was looking at creating a dashboard to display some information, the information required is found on an email. I decided to use Python to collect the data from the email and then convert to a HTML file to display. I am now looking to amend this to make it move visually pleaseing.